Converted to ARC for iOS Move
We had to go to ARC to get this code onto iOS, but at the same time, the
execution times went from 38 msec to 58 msec - a significant jump to get
rid of retain/release and use ARC. But there's no way to use retain/release
on iOS, so we have to go anyway. Just sad to see the times suffer.