Fixed up the int casting
In Xcode 9, there is a casting that needed to be done, and then I wanted
to add in the changes to the README so that the folks reading this would
be about to see what this was all about, and how it basically
functioned. Nothing great, but it was nice to check it out again.