Added all the param and return tags
The method documentation in the comments needed to have all the
parameters and return values commented to match the work we have done
with the Swift code, and this is just bring it up to par. I'm a little
disappointed that you can't use Markdown in the ObjC comments like you
can the Swift ones, but so it goes.