Updated the code to Xcode 10
Wow... I wanted to fix up the project for Xcode 10, and in doing that, I
found that when I timed the same routines in ObjC vs. Swift, I was
getting amazingly good performance in ObjC, and that just blew me away.
We are now getting sub-msec times on the solution once the words are
loaded. That's just wild.