Cleaned up ComboBox to PopUps
One of the issues with the old code was that the hint was a pair of
ComboBoxes, and I really liked the NSPopUpButton more in the Swift code.
Add to that the hint text on the text fields, and then the simplicity
that the test mode ran in Swift, and I decided to tackly all of them at
once. It was a good time - Xcode 10 is getting better and better.