Moved to Cypher Pattern matching
The previous code had a little unusual way of matching the pattterns for
the cyphertext and the plaintext possible solutions. This was a
hold-over from the C-version, and with the pattern-matching that came
from the Clojure code, it seemed like a good idea to put that matching
into the ObjC code as well. It was mostly done in C, with low-level
functions - because that was the simplest way to do it. It works like a
charm, and really doesn't impact performance, because this is all done
BEFORE the attack is started.