Moved from the lookup to simple math
There was no need to have the lookup of the character based on it's
place in the source string - we can just add the offset to the value of
'a', and we're good to go. This is really an advantage of ObjC, and you
can't do this in Swift - because they didn't allow for it in the
Character class. Anyway, this now handles the same thing without the
creation of the lookup table.